working on networking trait
