+++ p2p/src/authenticated/connection/stream.rs
